refactor: EventRecord 인터페이스 선언 및 상속하도록 변경
