+++ search-service/src/main/java/com/example/query/kafka/record/EventRecord.java
@@ -1,0 +1,5 @@
+package com.example.query.kafka.record;
+
+public interface EventRecord {
+    Long getPartitionKey();
+}
