+++ search-service/src/main/java/com/example/query/kafka/record/UpdateTransactionRecord.java
@@ -1,0 +1,8 @@
+package com.example.query.kafka.record;
+
+public record UpdateTransactionRecord(long id, long predictCost, boolean isReliable) implements EventRecord {
+    @Override
+    public Long getPartitionKey() {
+        return id;
+    }
+}
